CPLP-497: file internal renamings
