+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/CatenaX.NetworkServices.UserAdministration.Service.csproj
