+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/Properties/launchSettings.json
@@ -17,7 +17,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "CatenaX.NetworkServices.Invitation.Service": {
+    "CatenaX.NetworkServices.UserAdministration.Service": {
       "commandName": "Project",
       "launchBrowser": true,
       "launchUrl": "weatherforecast",
