+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/helmchart/Chart.yaml
@@ -1,5 +1,5 @@
 apiVersion: v1
 appVersion: "1.0"
 description: A Helm chart for Kubernetes
-name: invitation-service
+name: useradministration-service
 version: 0.1.0
