+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/helmchart/templates/ingres.yaml
@@ -1,7 +1,7 @@
 apiVersion: networking.k8s.io/v1
 kind: Ingress
 metadata:
-  name: "{{ template "invitation-service.fullname" . }}-ingress"
+  name: "{{ template "useradministration-service.fullname" . }}-ingress"
   annotations:
     cert-manager.io/cluster-issuer: letsencrypt-{{ .Values.ingress.clusterIssuer }}
     acme.cert-manager.io/http01-ingress-class: service
@@ -20,11 +20,11 @@ spec:
   - host: {{ .Values.ingress.serviceHost }}
     http:
       paths:
-      - path: /api/invitation(/|$)(.*)
+      - path: /api/useradministration(/|$)(.*)
         pathType: Prefix
         backend:
           service:
-            name: {{ template "invitation-service.name" . }}-service
+            name: {{ template "useradministration-service.name" . }}-service
             port:
               number: 8080
