+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/helmchart/templates/service.yaml
@@ -1,7 +1,7 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: {{ template "invitation-service.name" . }}-service
+  name: {{ template "useradministration-service.name" . }}-service
 spec:
   type: ClusterIP
   ports:
@@ -9,4 +9,4 @@ spec:
     protocol: "TCP"
     targetPort: 80
   selector:
-    app: {{ template "invitation-service.name" . }}
+    app: {{ template "useradministration-service.name" . }}
