+++ coreservices/useradministration/src/CatenaX.NetworkServices.Invitation.Service/helmchart/values.yaml
@@ -1,7 +1,7 @@
 # Default values for container-blueprint.
 # This is a YAML-formatted file.
 # Declare variables to be passed into your templates.
-fullnameOverride: invitation-service
+fullnameOverride: useradministration-service
 image:
   repository: test
   tag: stable
